fix api locations
